Replace Catch2 with doctest (#4)
* Replace Catch2 with doctest

Catch2 did not fall short, and I am not using the features that
distinguish doctest from Catch2 (i.e. writing tests in the same
source file) yet.
But it's good to have the option going forward.
Moreover, doctest can be easily installed on systems on which there's no
package manager, because it's a single file dependency.